cppcanvas: try to make MSVC tinderbox happy

+++ cppcanvas/source/mtfrenderer/emfplus.cxx
@@ -1343,12 +1343,12 @@ namespace cppcanvas
                             B2DSize mappedSize( MapSize (dw/2, dh/2));
 
                             float endAngle = startAngle + sweepAngle;
-                            startAngle = fmod (startAngle, M_PI*2);
+                            startAngle = fmodf(startAngle, M_PI*2);
                             if (startAngle < 0)
-                                startAngle += M_PI*2;
-                            endAngle = fmod (endAngle, M_PI*2);
+                                startAngle += static_cast<float>(M_PI*2);
+                            endAngle = fmodf(endAngle, M_PI*2);
                             if (endAngle < 0)
-                                endAngle += M_PI*2;
+                                endAngle += static_cast<float>(M_PI*2);
 
                             if (sweepAngle < 0)
                                 std::swap (endAngle, startAngle);
